+++ src/app/components/SettingsMenu/index.tsx
@@ -85,7 +85,7 @@ import {localize} from "../../services/utilities/localize";
 import './style.css';
 import 'react-image-crop/dist/ReactCrop.css';
 
-export const C_VERSION = '0.25.71';
+export const C_VERSION = '0.25.72';
 export const C_CUSTOM_BG_ID = 'river_custom_bg';
 
 export const languageList = [{